% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartoita_lista.R
\name{kartoita_lista}
\alias{kartoita_lista}
\title{Kartoita lista}
\usage{
kartoita_lista(
  kartoitettava_lista_fp,
  datrows = 10,
  datcols = 10,
  target_fp,
  color_vector = NA
)
}
\arguments{
\item{kartoitettava_lista_fp}{Polku, jossa listamuotoinen kartta on. Tiedostomuoto oltava csv.}

\item{datrows}{Tehtävän kartan rivimäärä. Oletuksena 10, mutta käytössä on myös 8 ja 12.}

\item{datcols}{Tehtävän kartan sarakemäärä. Oletuksena 10, mutta käytössä on myös 8 ja 12.}

\item{target_fp}{Polku, johon valmis kartta tallennetaan. Tiedostomuoto xlsx.}

\item{color_vector}{Dataframe, jossa sarakkeet 'naytenimi' ja 'color'. 'naytenimi' on jonkin näytteen nimi sellaisena kuin se kartalle merkitään. 'color' on openxlsx-yhteensopiva väri, esim 'red', tai hex-koodi.}
}
\description{
Funktio muuttaa listamuotoon kootun näytekartan takaisin karttamuotoon.
Argumentteja ovat kartoitettavan listan (.csv-muotoinen) tiedostopolku, rasioiden rivi- ja sarakemäärät,
tiedostopolku tallennettavalle kartalle, sekä valinnainen solujen värit määräävä taulu.
Tehtävälle kartalle lihavoidaan otsikkorivit, ja
}
