% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listoita_kartta.R
\name{listoita_kartta}
\alias{listoita_kartta}
\title{Listoita kartta}
\usage{
listoita_kartta(
  naytekarttalista_fp = NULL,
  naytekartat = NULL,
  export_path = "./Kohortti_kaikki_naytteet.csv",
  export_rds_path = "./Kohortti_kaikki_naytteet.Rds",
  write = TRUE,
  doreturn = FALSE
)
}
\arguments{
\item{naytekarttalista_fp}{Polku naytekarttalistalle, joka on excel-tiedosto missä on sarakkeet Aineisto ja Sijainti. Vaihtoehtoinen parametrin 'naytekartat' kanssa.}

\item{naytekartat}{Dataframe, jossa sarakkeet Aineisto ja Sijainti. Sisältö sama kuin 'naytekarttalista_fp' -parametrin exceltiedstossa. Vaihtoehtoinen parametrin 'naytekarttalista_fp' kanssa.}

\item{export_path}{Polku, johon .csv-muotoinen listoitettu kartta tallennetaan.}

\item{export_rds_path}{Polku, johon .Rds-muotoinen listoitettu kartta tallennetaan.}

\item{write}{TRUE/FALSE. Kirjoitetaanko csv ja rds-tiedostoja?}

\item{doreturn}{TRUE/FALSE. Palautetaanko koottu lista?}
}
\value{
Listoitettu kartta
}
\description{
Lukee xlsx-muotoisen näytekartan, ja muokkaa sen Poimuri-yhteensopivaksi listaksi.
Funktio käyttää apuna listoita_kartta_apufunktiot.R -tiedoston funktioita.
}
